pyproject.toml
[project]
name = "agv-ops-agent"
version = "1.0.0"
description = "AGV 智能运维助手 - 生产环境"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "fastapi-jwt[authlib]>=0.3.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true


[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "ipykernel==6.30.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.9",
]
